Chore/header file separation (#2)
* chore: Separated header file

* docs: documented header file

* fix: remove unused timer typedef and formatting
